Add files via upload
SQL stands for Structured Query Language. It is a programming language used for managing and manipulating relational databases. SQL allows users to create, modify, and query databases, as well as perform various operations such as inserting, updating, and deleting data. It is widely used in web development, data analysis, and other applications that involve working with databases. SQL is a standardized language, so it can be used with different database management systems like MySQL, Oracle, SQL Server, and PostgreSQL.